Basic structure of a declaration shadowing lint
This is a first attempt only, and does not cover all cases. The basic
idea is to use check_pat() method of LintPass to look up the given pat
in each enclosing scope.

The check_pat() method will need to eventually comply with Rust code
style and idioms.

check_pat() is also perhaps not the only method on LintPass that will be
needed for declaration shadowing detection to work.