A better attempt at shadow detection with Visitor
We define a Visitor which maintains a mapping of Names to
NodeIds. During analysis of Pats, if a Name is determined to exist in
the Visitor's mappings, we check if the pre-existing Name encloses the
current Name which indicates a shadow. We emit a lint and a span_note
describing the shadowed name and the locations of the shadowing names.