More Guide editing, up to Strings section
More work on rust-lang/rust#19335

Changes up to but not including the Strings section. Apart from
rewording some unclear, oddly-phrased, or slightly incorrect passages,
the most significant changes are the removal of short sections of code
or passages which are more appropriate for rookies than programmers who
are merely new to systems programming with Rust.